requirements.txt
asgiref==3.9.1
crispy-bootstrap5==2025.6
Django==5.2.4
django-ckeditor==6.7.3
django-ckeditor-5==0.2.18
django-crispy-forms==2.4
django-js-asset==3.1.2
gunicorn==23.0.0
packaging==25.0
pillow==11.3.0
sqlparse==0.5.3
tzdata==2025.2
